Improve the functioning of Navbar Buttons
This commit will solve the problem of navbar buttons
getting hover effect even when they are not clickable,
which was making it a bit confusing and irritating for
the user. I have basically removed the classes and ids
which were given to `<li>`, and gave them to the anchor
`<a>` tags inside the respective `<li>` tags, which
changes only one thing, that the hover effect and active
state changes are applied to `<a>` rather than `<li>`,
which makes it clickable and thus improve the functionality
of the navbar.